# cmd/core/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "quebec core 模块后端",
        "title": "quebec core Swagger API接口文档",
        "contact": {
            "name": "Lyon Mu",
            "url": "https://github.com/lyonmu",
            "email": "lyonmu@foxmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/lyonmu/quebec/blob/master/LICENSE"
        },
        "version": "v0.0.1"
    },
    "host": "http://localhost:59024",
    "basePath": "/core/api",
    "paths": {
...
        "/v1/system/onlineuser/clearance/{id}": {
            "delete": {
                "description": "清除在线用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "清除在线用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/system/onlineuser/list": {
            "get": {
                "description": "获取在线用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "在线用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问IP",
                        "name": "access_ip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 1000,
                        "minimum": 10,
                        "type": "integer",
                        "default": 10,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "50000,success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/response.SystemOnlineUserListResp"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "code.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 50000
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "common.OperationType": {
            "type": "integer",
            "enum": [
                1
            ],
            "x-enum-varnames": [
                "OperationLogin"
            ]
        },
...
        "response.SystemOnlineUserListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "在线用户列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SystemOnlineUserResp"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "response.SystemOnlineUserResp": {
            "type": "object",
            "properties": {
                "access_ip": {
                    "description": "访问IP",
                    "type": "string"
                },
                "browser_engine_name": {
                    "description": "浏览器引擎名称",
                    "type": "string"
                },
                "browser_engine_version": {
                    "description": "浏览器引擎版本",
                    "type": "string"
                },
                "browser_name": {
                    "description": "浏览器名称",
                    "type": "string"
                },
                "browser_version": {
                    "description": "浏览器版本",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "last_operation_time": {
                    "description": "最后操作时间",
                    "type": "integer"
                },
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "operation_type": {
                    "description": "操作类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.OperationType"
                        }
                    ]
                },
                "os": {
                    "description": "操作系统",
                    "type": "string"
                },
                "platform": {
                    "description": "操作平台",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "token认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}