# cmd/core/internal/dto/response/system.go
package response

import (
	"github.com/lyonmu/quebec/cmd/core/internal/common"
	"github.com/lyonmu/quebec/cmd/core/internal/ent"
	"github.com/lyonmu/quebec/pkg/constant"
)

type CaptchaResponse struct {
	Id       string `json:"id" example:"1234567890"`   // 验证码id
	Pictures string `json:"pictures" example:"base64"` // 验证码图片
	Length   int    `json:"length" example:"4"`        // 验证码长度
}

type SystemInfoResponse struct {
	Username string `json:"username"`  // 用户名
	Nickname string `json:"nickname"`  // 用户昵称
	Token    string `json:"token"`     // token
	RoleName string `json:"role_name"` // 角色名称
}

type SystemOnlineUserResp struct {
	ID                   string               `json:"id"`                               // ID
	Username             string               `json:"username"`                         // 用户名
	Nickname             string               `json:"nickname"`                         // 用户昵称
	AccessIP             string               `json:"access_ip"`                        // 访问IP
	LastOperationTime    int64                `json:"last_operation_time"`              // 最后操作时间
	OperationType        common.OperationType `json:"operation_type"`                   // 操作类型
	Os                   string               `json:"os,omitempty"`                     // 操作系统
	Platform             string               `json:"platform,omitempty"`               // 操作平台
	BrowserName          string               `json:"browser_name,omitempty"`           // 浏览器名称
	BrowserVersion       string               `json:"browser_version,omitempty"`        // 浏览器版本
	BrowserEngineName    string               `json:"browser_engine_name,omitempty"`    // 浏览器引擎名称
	BrowserEngineVersion string               `json:"browser_engine_version,omitempty"` // 浏览器引擎版本
}

func (r *SystemOnlineUserResp) LoadDb(e *ent.CoreOnLineUser) {
	r.ID = e.UserID
	r.AccessIP = e.AccessIP
	r.LastOperationTime = e.LastOperationTime
	r.OperationType = e.OperationType
	r.Os = e.Os
	r.Platform = e.Platform
	r.BrowserName = e.BrowserName
	r.BrowserVersion = e.BrowserVersion
	r.BrowserEngineName = e.BrowserEngineName
	r.BrowserEngineVersion = e.BrowserEngineVersion

	if e.Edges.OnLineFromUser != nil {
		r.Username = e.Edges.OnLineFromUser.Username
		r.Nickname = e.Edges.OnLineFromUser.Nickname
	}
}

type SystemOnlineUserListResp struct {
	Total    int                     `json:"total,omitempty"`     // 总条数
	Items    []*SystemOnlineUserResp `json:"items,omitempty"`     // 在线用户列表
	Page     int                     `json:"page,omitempty"`      // 页码
	PageSize int                     `json:"page_size,omitempty"` // 每页条数
}

type SystemRoleResp struct {
	ID     string           `json:"id,omitempty"`     // 角色ID
	Name   string           `json:"name,omitempty"`   // 角色名称
	Remark string           `json:"remark,omitempty"` // 角色备注
	Status constant.YesOrNo `json:"status,omitempty"` // 角色状态 [1: 启用, 2: 禁用]
}

func (r *SystemRoleResp) LoadDb(e *ent.CoreRole) {
	r.ID = e.ID
	r.Name = e.Name
	r.Remark = e.Remark
	r.Status = e.Status
}

type SystemRoleListResp struct {
	Total    int               `json:"total,omitempty"`     // 总条数
	Items    []*SystemRoleResp `json:"items,omitempty"`     // 角色列表
	Page     int               `json:"page,omitempty"`      // 页码
	PageSize int               `json:"page_size,omitempty"` // 每页条数
}

type SystemUserResp struct {
	ID                 string               `json:"id,omitempty"`                   // 用户ID
	Username           string               `json:"username,omitempty"`             // 用户名
	Nickname           string               `json:"nickname,omitempty"`             // 用户昵称
	Email              string               `json:"email,omitempty"`                // 用户邮箱
	Status             constant.YesOrNo     `json:"status,omitempty"`               // 用户状态 [1: 启用, 2: 禁用]
	RoleID             string               `json:"role_id,omitempty"`              // 角色ID
	RoleName           string               `json:"role_name,omitempty"`            // 角色名称
	LastPasswordChange int64                `json:"last_password_change,omitempty"` // 最后密码修改时间
	LastOperationTime  int64                `json:"last_operation_time,omitempty"`  // 最后操作时间
	OperationType      common.OperationType `json:"operation_type,omitempty"`       // 操作类型
	Remark             string               `json:"remark,omitempty"`               // 用户备注
}

func (r *SystemUserResp) LoadDb(e *ent.CoreUser) {
	r.ID = e.ID
	r.Username = e.Username
	r.Nickname = e.Nickname
	r.Email = e.Email
	r.Status = e.Status
	r.LastPasswordChange = e.LastPasswordChange
	r.Remark = e.Remark

	if e.Edges.UserFromRole != nil {
		r.RoleName = e.Edges.UserFromRole.Name
		r.RoleID = e.Edges.UserFromRole.ID
	}
	if e.Edges.OnLineToUser != nil {
		if len(e.Edges.OnLineToUser) > 0 {
			r.LastOperationTime = e.Edges.OnLineToUser[0].LastOperationTime
			r.OperationType = e.Edges.OnLineToUser[0].OperationType
		}
	}
}

type SystemUserListResp struct {
	Total    int               `json:"total,omitempty"`     // 总条数
	Items    []*SystemUserResp `json:"items,omitempty"`     // 用户列表
	Page     int               `json:"page,omitempty"`      // 页码
	PageSize int               `json:"page_size,omitempty"` // 每页条数
}